Send photo attachments from RP on VK channel
